soylent_red_division/pyproject.toml
[project]
name = "soylent_red_division"
version = "0.1.0"
description = "soylent_red_division using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "crewai-tools",
    "requests",
    "beautifulsoup4",
    "PyYAML"
]

[project.scripts]
soylent_red_division = "soylent_red_division.main:run"
run_crew = "soylent_red_division.main:run"
train = "soylent_red_division.main:train"
replay = "soylent_red_division.main:replay"
test = "soylent_red_division.main:test"
brand_author_draft = "soylent_red_division.main:brand_author_draft"
feedback = "soylent_red_division.main:feedback"
signoff = "soylent_red_division.main:signoff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"